+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsNetworkIF.java
@@ -51,6 +51,7 @@ public final class WindowsNetworkIF extends AbstractNetworkIF {
 
     private static final boolean IS_VISTA_OR_GREATER = VersionHelpers.IsWindowsVistaOrGreater();
     private static final byte CONNECTOR_PRESENT_BIT = 0b00000100;
+    private static final int MIB_IF_TYPE_LOOPBACK = 24;
 
     private int ifType;
     private int ndisPhysicalMediumType;
@@ -68,9 +69,9 @@ public final class WindowsNetworkIF extends AbstractNetworkIF {
     private String ifAlias;
     private IfOperStatus ifOperStatus;
 
-    public WindowsNetworkIF(NetworkInterface netint) throws InstantiationException {
+    public WindowsNetworkIF(NetworkInterface netint, MIB_IF_ROW2 ifRow) throws InstantiationException {
         super(netint);
-        updateAttributes();
+        updateAttributes(ifRow);
     }
 
     /**
@@ -82,9 +83,12 @@ public final class WindowsNetworkIF extends AbstractNetworkIF {
      */
     public static List<NetworkIF> getNetworks(boolean includeLocalInterfaces) {
         List<NetworkIF> ifList = new ArrayList<>();
-        for (NetworkInterface ni : getNetworkInterfaces(includeLocalInterfaces)) {
+        for (NetworkInterface ni : getNetworkInterfaces(true)) {
             try {
-                ifList.add(new WindowsNetworkIF(ni));
+                MIB_IF_ROW2 ifRow = queryIfEntry(ni.getIndex());
+                if (includeLocalInterfaces || ifRow != null && ifRow.Type != MIB_IF_TYPE_LOOPBACK) {
+                    ifList.add(new WindowsNetworkIF(ni, ifRow));
+                }
             } catch (InstantiationException e) {
                 LOG.debug("Network Interface Instantiation failed: {}", e.getMessage());
             }
@@ -169,56 +173,80 @@ public final class WindowsNetworkIF extends AbstractNetworkIF {
 
     @Override
     public boolean updateAttributes() {
-        // MIB_IFROW2 requires Vista (6.0) or later.
-        if (IS_VISTA_OR_GREATER) {
-            // Create new MIB_IFROW2 and set index to this interface index
-            MIB_IF_ROW2 ifRow = new MIB_IF_ROW2();
-            ifRow.InterfaceIndex = queryNetworkInterface().getIndex();
-            if (0 != IPHlpAPI.INSTANCE.GetIfEntry2(ifRow)) {
-                // Error, abort
-                LOG.error("Failed to retrieve data for interface {}, {}", queryNetworkInterface().getIndex(),
-                        getName());
-                return false;
-            }
-            this.ifType = ifRow.Type;
-            this.ndisPhysicalMediumType = ifRow.PhysicalMediumType;
-            this.connectorPresent = (ifRow.InterfaceAndOperStatusFlags & CONNECTOR_PRESENT_BIT) > 0;
-            this.bytesSent = ifRow.OutOctets;
-            this.bytesRecv = ifRow.InOctets;
-            this.packetsSent = ifRow.OutUcastPkts;
-            this.packetsRecv = ifRow.InUcastPkts;
-            this.outErrors = ifRow.OutErrors;
-            this.inErrors = ifRow.InErrors;
-            this.collisions = ifRow.OutDiscards; // closest proxy
-            this.inDrops = ifRow.InDiscards; // closest proxy
-            this.speed = ifRow.ReceiveLinkSpeed;
-            this.ifAlias = Native.toString(ifRow.Alias);
-            this.ifOperStatus = IfOperStatus.byValue(ifRow.OperStatus);
-        } else {
-            // Create new MIB_IFROW and set index to this interface index
-            MIB_IFROW ifRow = new MIB_IFROW();
-            ifRow.dwIndex = queryNetworkInterface().getIndex();
-            if (0 != IPHlpAPI.INSTANCE.GetIfEntry(ifRow)) {
-                // Error, abort
-                LOG.error("Failed to retrieve data for interface {}, {}", queryNetworkInterface().getIndex(),
-                        getName());
-                return false;
-            }
-            this.ifType = ifRow.dwType;
-            // These are unsigned ints. Widen them to longs.
-            this.bytesSent = ParseUtil.unsignedIntToLong(ifRow.dwOutOctets);
-            this.bytesRecv = ParseUtil.unsignedIntToLong(ifRow.dwInOctets);
-            this.packetsSent = ParseUtil.unsignedIntToLong(ifRow.dwOutUcastPkts);
-            this.packetsRecv = ParseUtil.unsignedIntToLong(ifRow.dwInUcastPkts);
-            this.outErrors = ParseUtil.unsignedIntToLong(ifRow.dwOutErrors);
-            this.inErrors = ParseUtil.unsignedIntToLong(ifRow.dwInErrors);
-            this.collisions = ParseUtil.unsignedIntToLong(ifRow.dwOutDiscards); // closest proxy
-            this.inDrops = ParseUtil.unsignedIntToLong(ifRow.dwInDiscards); // closest proxy
-            this.speed = ParseUtil.unsignedIntToLong(ifRow.dwSpeed);
-            this.ifAlias = ""; // not supported by MIB_IFROW
-            this.ifOperStatus = IfOperStatus.UNKNOWN; // not supported
+        MIB_IF_ROW2 ifRow = queryIfEntry(queryNetworkInterface().getIndex());
+        return updateAttributes(ifRow);
+    }
+
+    private boolean updateAttributes(MIB_IF_ROW2 ifRow) {
+        if (ifRow == null) {
+            LOG.error("Failed to retrieve data for interface {}, {}", queryNetworkInterface().getIndex(), getName());
+            return false;
         }
+        this.ifType = ifRow.Type;
+        this.ndisPhysicalMediumType = ifRow.PhysicalMediumType;
+        this.connectorPresent = (ifRow.InterfaceAndOperStatusFlags & CONNECTOR_PRESENT_BIT) > 0;
+        this.bytesSent = ifRow.OutOctets;
+        this.bytesRecv = ifRow.InOctets;
+        this.packetsSent = ifRow.OutUcastPkts;
+        this.packetsRecv = ifRow.InUcastPkts;
+        this.outErrors = ifRow.OutErrors;
+        this.inErrors = ifRow.InErrors;
+        this.collisions = ifRow.OutDiscards; // closest proxy
+        this.inDrops = ifRow.InDiscards; // closest proxy
+        this.speed = ifRow.ReceiveLinkSpeed;
+        this.ifAlias = Native.toString(ifRow.Alias);
+        this.ifOperStatus = IfOperStatus.byValue(ifRow.OperStatus);
         this.timeStamp = System.currentTimeMillis();
         return true;
+    }
+
+    private static MIB_IF_ROW2 queryIfEntry(int index) {
+        // Create new MIB_IFROW2 and set index to this interface index
+        MIB_IF_ROW2 ifRow = new MIB_IF_ROW2();
+        if (IS_VISTA_OR_GREATER) {
+            // Set index to this interface index
+            ifRow.InterfaceIndex = index;
+            if (0 == IPHlpAPI.INSTANCE.GetIfEntry2(ifRow)) {
+                return ifRow;
+            }
+        } else {
+            // MIB_IFROW2 requires Vista (6.0) or later. Use MIB_IFROW and copy over
+            MIB_IFROW ifRowXP = new MIB_IFROW();
+            // Set index to this interface index
+            ifRowXP.dwIndex = index;
+            if (0 == IPHlpAPI.INSTANCE.GetIfEntry(ifRowXP)) {
+                ifRow.Type = ifRowXP.dwType;
+                ifRow.OutOctets = ParseUtil.unsignedIntToLong(ifRowXP.dwOutOctets);
+                ifRow.InOctets = ParseUtil.unsignedIntToLong(ifRowXP.dwInOctets);
+                ifRow.OutUcastPkts = ParseUtil.unsignedIntToLong(ifRowXP.dwOutUcastPkts);
+                ifRow.InUcastPkts = ParseUtil.unsignedIntToLong(ifRowXP.dwInUcastPkts);
+                ifRow.OutErrors = ParseUtil.unsignedIntToLong(ifRowXP.dwOutErrors);
+                ifRow.InErrors = ParseUtil.unsignedIntToLong(ifRowXP.dwInErrors);
+                ifRow.OutDiscards = ParseUtil.unsignedIntToLong(ifRowXP.dwOutDiscards); // closest proxy
+                ifRow.InDiscards = ParseUtil.unsignedIntToLong(ifRowXP.dwInDiscards); // closest proxy
+                ifRow.ReceiveLinkSpeed = ParseUtil.unsignedIntToLong(ifRowXP.dwSpeed);
+                ifRow.Alias = new char[0];
+                switch (ifRowXP.dwOperStatus) {
+                case 4: // IF_OPER_STATUS_CONNECTED
+                case 5: // IF_OPER_STATUS_OPERATIONAL
+                    ifRow.OperStatus = 1; // Up
+                    ifRow.InterfaceAndOperStatusFlags = CONNECTOR_PRESENT_BIT;
+                    break;
+                case 0: // IF_OPER_STATUS_NON_OPERATIONAL
+                case 3: // IF_OPER_STATUS_CONNECTING
+                    ifRow.InterfaceAndOperStatusFlags = CONNECTOR_PRESENT_BIT;
+                    ifRow.OperStatus = 2; // Down
+                    break;
+                case 1: // IF_OPER_STATUS_UNREACHABLE
+                case 2: // IF_OPER_STATUS_DISCONNECTED
+                    ifRow.OperStatus = 2; // Down
+                    break;
+                default:
+                    ifRow.OperStatus = 4; // Unknown
+                }
+                // PhysicalMediumType = 0, Unspecified
+            }
+        }
+        return null;
     }
 }