+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ##### Bug fixes / Improvements
 * [#1596](https://github.com/oshi/oshi/pull/1596): Single COM initialization for groups of queries - [@dbwiddis](https://github.com/dbwiddis).
 * [#1603](https://github.com/oshi/oshi/pull/1603): Improve performance of Windows USB device tree parsing - [@dbwiddis](https://github.com/dbwiddis).
+* [#1607](https://github.com/oshi/oshi/pull/1607): More efficiently test for local interfaces - [@dbwiddis](https://github.com/dbwiddis).
+
 
 # 5.7.0 (2021-04-01)